Started The Porting Process for BigNum Class Library
I looked up Nick Gammon's arduino library port for the java "bc_class"
and decided to work arond the SRAM limitations of bignum.  The ultimate
plan is to free the numbers immediately after processing so our sram
heap wont overflow andn crash. Many lines have changed and isnt a
compilable build YET. Everything is disorganised to a large degree. Go
update the handling and printing of the numbers as a planned fix.