Fixed Truncated Output Of Total
Finally fixed it, it required only a simple fix. By adding an "isTotal"
boolean before printing any LCD parameter, one can easily bypass the
check on line 541 of this commit version and make sure the decimal
carries over. Not only does it carry over on printing the total result,
but works as well for result_as_input(). It was a perfect fix.