subroutine conditioning changesq
I changed a couple things due to the verbosity and redundancy of certain subroutine checks.
I had two identical if else statements doing the same action of stacking the operator one above ( p + 1 )
also both of them checked to see if it was equal to 6 and not equal to 6 and if the selected operator was less than, greater than and equal to.

Point is, both sides of the conditions if true and if not true led to the same operator action. != 6 && = 6   and...  operator > opstack[p] && operator <= opstack[p]  which led to pushing to top of stack eitherway, so its redundant. (It works, but is redundant and confusing)
*untested on an actual arduino*