Working Build 3.0 POSTFIX ALGORITHM FINALIZED
I was originally overthinking the solution to this problem. While I was originally focused on changing and condensing the for,if-else conditionals, the fix of the solution was rather simple. Since 8 is the handling ID for the exponent ^, since it was equal to the currently selected operator, it wouldn't get copied over.
Code:
else if ((opr8tr != 7 && opr8tr > postfix_opstack[p] && postfix_opstack[p] != 0 && postfix_opstack[p] != 6))  {

once 8 hits and its not greater than 8, then it wouldnt get copied to the stack in the first place, so I changed the symbol
> to >= and now it copies normally.

else if ((opr8tr != 7 && opr8tr >= postfix_opstack[p] && postfix_opstack[p] != 0 && postfix_opstack[p] != 6))  {

and now the postfix algorithm for handling exponents works perfectly.