Started Postfix Evaluator and Additional Logic
Implemented basic logic for evaluating the reference stack. First, start from the 0 position of the postfix_reference_stack and loop till the end of the stack. Compare the values ( if it's 1, the # indicator, or if it's anything else, which would be an operator )

Value = postfix_stack_reference @ [p]
When value is  1, increment the number_stack_index by 1.
If its anything above one, then perform_operation with that value of stack reference while giving the position of where that operator was found into the next function.

perform_operation( which_operator , position_of_operator_in_the_reference_stack );

Under perform_operation(), we use the given position and move backwards to the beginning of the reference stack.
 
PLAN MOVING FORWARD: We count the # of 1's and make sure its the first 2 behind the operator. Then figure out their respective index location by counting forward again. Then perform operation between the 2 numbers by calling the indices from the NUMBERSTACK FINAL and using the operator selected.