Edited Structure of the perform_operation.
perform_operation(4, 3)  operator, xpos
then go back to the beginning of the stack and then check for #1 code indicators till the end. He total # of 1's collected to the end of the stack means the index of the number stack to perform the calculation on.
Doesn't matter the order that you count the number index because it still gives the proper distance.
number index - 1 (do operation) number index = gives result.'

Improve the rest of the handling but this is how we do the rest of the code.