updated debug
new method for debugging and also improved diagnostics of the problem. It is related to the reference stack not deleting the 1's that are used up. So extra 1's will be counted and thus start to perform numbers on 0-value indicies. pop hte reference stack as well as the numberstack (or at least delete the 1's in the reference stack as they are used. sad thing is though, we tamper with the reference stack :(