Improved handling of numberstack
The problem as stated in the previous commit, showed that the reference stack didnt reflect the changes that the numberstack had. (In terms of the amount of operands present.) When the numbers in the numberstack cascaded when various operations were performed, the reference stack kept counting numbers from the # of 1's present in it.

111411253 postfix reference stack
3 4 2 1 5 numberstack

4 = * multiplication on numberstack[1] and numberstack[2] which gets us

3 8 1 5 numberstack
11{1}411253 ( it doesnt remove the extra 1 right before the 4  every concurrent operation will be looking at an index position higher than the number it is supposed to hit )

by using the delete_ones variable, we can correct that offset every time we do an operation and increments with further operations. thus eliminating the chance of having errors.

UNTESTED but theoretically this fix is better than adding a second postfix stack and destroying its integrity.