Debug Improved, and Peripheral Logic Revised
Tweaked the Status LED code for more flexibility and added the special line. Cna be used for shift or other modifiers. Changed the attiny84keypadserial so it doesn't use the debounce code. (Redundant and wastes 2 PWM pins) Revised the Debug Serial for my other MEGA to more easily verify communication between the chips.