Structure and Error Handling Improvement AGAIN Part 2
Finalized error detection of double operators such as ' ++ , --, ^+, *- ' as well as '+)  ,  (^' these instances are the only cases where most of the double operator errors should occure. Calculator is 75% optimized. The structure has to be rewritten for some functions like pushtostack() and evaluate_postfix(). Aside from that, this is shippable as a consumer product at this design stage. (Most CompSci experts could disagree :P )