Finished New Graphing Implement For Real This Time
Fixed all the graphing,  and memory managemetn issues and now code works perfectly. One problem, the display does not like handling bigNumbers so we would need to invent a cropping mechanism of the sorts. Detect when theres outof bounds behavior when drawing the graph and stop it. Verbose for, prevent the given Adafruit Library from being overloaded with arbitrarily large numbers. (Something that the ti84 does as well)